// src/types/objective-types.ts
export type HelpBadge = {
  path: string;
  content: string;
  color: string;
};

export enum ObjectiveType {
  POLICY_CREATION_OBJECTIVE = 'POLICY_CREATION_OBJECTIVE',
  RESOURCE_POLICY_CREATION_OBJECTIVE = 'RESOURCE_POLICY_CREATION_OBJECTIVE',
  SCP_CREATION_OBJECTIVE = 'SCP_CREATION_OBJECTIVE',
  POLICY_EDIT_OBJECTIVE = 'POLICY_EDIT_OBJECTIVE',
  TRUST_POLICY_EDIT_OBJECTIVE = 'TRUST_POLICY_EDIT_OBJECTIVE',
  ROLE_CREATION_OBJECTIVE = 'ROLE_CREATION_OBJECTIVE',
  IAM_USER_GROUP_CREATION_OBJECTIVE = 'IAM_USER_GROUP_CREATION_OBJECTIVE',
  EDGE_CONNECTION_OBJECTIVE = 'EDGE_CONNECTION_OBJECTIVE',
  LEVEL_OBJECTIVE = 'LEVEL_OBJECTIVE',
  PERMISSION_BOUNDARY_CREATION_OBJECTIVE = 'PERMISSION_BOUNDARY_CREATION_OBJECTIVE',
}

export type BaseFinishEventMap = Record<ObjectiveType, string>;
// ...
export interface IAMPolicyEditObjective<
  TFinishEventMap extends BaseFinishEventMap,
  TValidateFn = string,
> {
  readonly id: string;
  /**
   * Unique identifier for the objective.
   * In creation objectives, the ID of the objective itself is sufficient to identify the validate function to use.
   * In edit objectives, we use `validate_fn_name` since we can have multiple edit objectives for the same node.
   */
  readonly validate_fn_name: TValidateFn;
  readonly type: ObjectiveType.POLICY_EDIT_OBJECTIVE;
  readonly entity: IAMCodeDefinedEntity;
  readonly allow_new_lines?: boolean;

  /**
   * Optional callout_message for the IAM Policy/Role Edit Objective.
   * Used to help the user understand what they need to do when editing the IAM Policy/Role.
   */
  readonly callout_message?: string;

  readonly on_finish_event: TFinishEventMap[ObjectiveType.POLICY_EDIT_OBJECTIVE];

  /**
   * Resources to grant to the users/groups associated with the IAM Policy/Role.
   */
  readonly resources_to_grant: PolicyGrantedAccess[];

  readonly help_badges?: HelpBadge[];
  readonly limit_new_lines?: boolean;
  readonly hint_messages?: { title: string; content: string }[];
  finished: boolean;
}